[legacy] move engine to legacy
